Broke CartesianProductFunction.pack() into packBinary(), packUnary(), and packLexical().
Fixed a few unit tests.

+++ java/src/edu/ohsu/cslu/parser/SortAndScanCsrSpmvParser.java
@@ -144,7 +144,7 @@ public class SortAndScanCsrSpmvParser extends
                             * rightCellRightChildren + rightChildrenProcessed;
 
                     tmpCartesianProductChildren[cartesianProductIndex] = grammar.cartesianProductFunction()
-                        .pack(leftChildIndex, (short) nonTerminalIndices[j]);
+                        .pack(leftChildIndex, nonTerminalIndices[j]);
                     tmpCartesianProductInsideProbabilities[cartesianProductIndex] = leftProbability
                             + insideProbabilities[j];
                     tmpCartesianProductMidpoints[cartesianProductIndex] = midpoint;
